+++ NEWS.md
@@ -5,6 +5,9 @@
 - New `get_usage()` function returns content usage data from Connect's `GET
   v1/instrumentation/content/hits` endpoint on Connect v2025.04.0 and higher.
   (#390)
+- New `get_integrations()` function lists all OAuth integrations available on the
+  Connect server from the `GET v1/oauth/integrations` endpoint on Connect v2024.12.0
+  and higher. (#413)
 
 ## Enhancements and fixes