+++ NEWS.md
@@ -1,4 +1,4 @@
-# connectapi (development version)
+# connectapi 0.8.0
 
 ## Breaking changes